Moved helper methods to CurrentBehavior.
