+++ src/DrillSergeant/BaseStep.cs
@@ -64,48 +64,7 @@ public abstract class BaseStep : IStep
     protected virtual void Dispose(bool disposing)
     {
     }
-
-    internal static object?[] ResolveParameters(IDictionary<string, object?> context, IDictionary<string, object?> input, ParameterInfo[] parameters)
-    {
-        var result = new object[parameters.Length];
-        var caster = new ReflectionParameterCaster();
-
-        if (result.Length > 0)
-        {
-            result[0] = caster.Cast(context, parameters[0].ParameterType);
-        }
-
-        if (result.Length > 1)
-        {
-            result[1] = caster.Cast(CopyInput(input), parameters[1].ParameterType);
-        }
-
-        return result;
-    }
-
-    internal static IDictionary<string, object?> CopyInput(IDictionary<string, object?> input)
-    {
-        var copy = new ExpandoObject();
-        var copyAsDict = (IDictionary<string, object?>)copy;
-
-        foreach (var (key, value) in input)
-        {
-            copyAsDict[key] = value;
-        }
-
-        return copy;
-    }
-
-    internal static void UpdateContext(IDictionary<string, object?> context, object changedContext)
-    {
-        var flags = BindingFlags.Public | BindingFlags.Instance | BindingFlags.GetProperty;
-
-        foreach (var property in changedContext.GetType().GetProperties(flags))
-        {
-            context[property.Name] = property.GetValue(changedContext);
-        }
-    }
-
+    
     internal static bool IsAsync(MethodInfo method) =>
         method.ReturnType.Name == nameof(Task) ||
         method.ReturnType.Name == typeof(Task<>).Name;
